backend/api/requirements.txt
fastapi
uvicorn
pypdf
google-generativeai
python-dotenv
python-multipart